dev: scratch crud and querying
