+++ src/main/kotlin/org/openmbee/flexo/mms/server/Routing.kt
@@ -99,6 +99,9 @@ fun Application.configureRouting() {
             metadataArtifacts()
             queryArtifacts()
 
+            crudScratch()
+            queryScratch()
+
             crudModel()
 
             crudCollections()
